refactor : Test Code Refactoring
