+++ src/test/java/usw/suwiki/controller/confirmationtoken/ConfirmationTokenControllerTestBase.java
@@ -3,12 +3,9 @@ package usw.suwiki.controller.confirmationtoken;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.datasource.init.ScriptUtils;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
-import usw.suwiki.BaseIntegrationTest;
+import usw.suwiki.IntegrationTestBase;
 import usw.suwiki.domain.confirmationtoken.ConfirmationToken;
 import usw.suwiki.domain.confirmationtoken.service.ConfirmationTokenCRUDService;
 
@@ -17,10 +14,7 @@ import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.List;
 
-@SpringBootTest
-@ActiveProfiles("test")
-@Transactional
-public class ConfirmationTokenControllerTest extends BaseIntegrationTest {
+public class ConfirmationTokenControllerTestBase extends IntegrationTestBase {
 
     @Autowired
     ConfirmationTokenCRUDService confirmationTokenCRUDService;
